MoveAllClipboard related tests to a new class
