+++ src/System.Windows.Forms/tests/UnitTests/TextBoxBaseTests.cs
@@ -4131,136 +4131,6 @@ public partial class TextBoxBaseTests
         Assert.Equal(0, createdCallCount);
     }
 
-    [ActiveIssue("https://github.com/dotnet/winforms/issues/11497")]
-    [WinFormsFact]
-    [SkipOnArchitecture(TestArchitectures.X86,
-        "Flaky tests, see: https://github.com/dotnet/winforms/issues/11497")]
-    public void TextBoxBase_ClearUndo_CanUndo_Success()
-    {
-        using SubTextBox control = new()
-        {
-            Text = "abc",
-            SelectionStart = 1,
-            SelectionLength = 2
-        };
-        control.Copy();
-
-        control.Text = "text";
-        control.SelectionLength = 2;
-        control.Paste();
-        Assert.Equal("bcxt", control.Text);
-
-        control.ClearUndo();
-        control.Undo();
-        Assert.Equal("bcxt", control.Text);
-    }
-
-    [WinFormsFact]
-    public void TextBoxBase_Copy_PasteEmpty_Success()
-    {
-        using SubTextBox control = new();
-        control.Copy();
-        Assert.Empty(control.Text);
-        Assert.True(control.IsHandleCreated);
-
-        control.Text = "text";
-        control.SelectionLength = 2;
-        Assert.Equal("text", control.Text);
-        Assert.True(control.IsHandleCreated);
-    }
-
-    [ActiveIssue("https://github.com/dotnet/winforms/issues/11558")]
-    [WinFormsFact]
-    [SkipOnArchitecture(TestArchitectures.X86,
-        "Flaky tests, see: https://github.com/dotnet/winforms/issues/11558")]
-    public void TextBoxBase_Copy_PasteNotEmpty_Success()
-    {
-        using SubTextBox control = new()
-        {
-            Text = "abc",
-            SelectionStart = 1,
-            SelectionLength = 2
-        };
-        control.Copy();
-        Assert.Equal("abc", control.Text);
-        Assert.True(control.IsHandleCreated);
-
-        control.Text = "text";
-        control.SelectionLength = 2;
-        control.Paste();
-        Assert.Equal("bcxt", control.Text);
-        Assert.True(control.CanUndo);
-        Assert.True(control.Modified);
-        Assert.True(control.IsHandleCreated);
-    }
-
-    [WinFormsFact]
-    public void TextBoxBase_Copy_PasteEmptyWithHandle_Success()
-    {
-        using SubTextBox control = new();
-        Assert.NotEqual(IntPtr.Zero, control.Handle);
-        int invalidatedCallCount = 0;
-        control.Invalidated += (sender, e) => invalidatedCallCount++;
-        int styleChangedCallCount = 0;
-        control.StyleChanged += (sender, e) => styleChangedCallCount++;
-        int createdCallCount = 0;
-        control.HandleCreated += (sender, e) => createdCallCount++;
-
-        control.Copy();
-        Assert.Empty(control.Text);
-        Assert.True(control.IsHandleCreated);
-        Assert.Equal(0, invalidatedCallCount);
-        Assert.Equal(0, styleChangedCallCount);
-        Assert.Equal(0, createdCallCount);
-
-        control.Text = "text";
-        control.SelectionLength = 2;
-        Assert.Equal("text", control.Text);
-        Assert.True(control.IsHandleCreated);
-        Assert.Equal(0, invalidatedCallCount);
-        Assert.Equal(0, styleChangedCallCount);
-        Assert.Equal(0, createdCallCount);
-    }
-
-    [ActiveIssue("https://github.com/dotnet/winforms/issues/11498")]
-    [WinFormsFact]
-    [SkipOnArchitecture(TestArchitectures.X86,
-        "Flaky tests, see: https://github.com/dotnet/winforms/issues/11498")]
-    public void TextBoxBase_Copy_PasteNotEmptyWithHandle_Success()
-    {
-        using SubTextBox control = new()
-        {
-            Text = "abc",
-            SelectionStart = 1,
-            SelectionLength = 2
-        };
-        Assert.NotEqual(IntPtr.Zero, control.Handle);
-        int invalidatedCallCount = 0;
-        control.Invalidated += (sender, e) => invalidatedCallCount++;
-        int styleChangedCallCount = 0;
-        control.StyleChanged += (sender, e) => styleChangedCallCount++;
-        int createdCallCount = 0;
-        control.HandleCreated += (sender, e) => createdCallCount++;
-
-        control.Copy();
-        Assert.Equal("abc", control.Text);
-        Assert.True(control.IsHandleCreated);
-        Assert.Equal(0, invalidatedCallCount);
-        Assert.Equal(0, styleChangedCallCount);
-        Assert.Equal(0, createdCallCount);
-
-        control.Text = "text";
-        control.SelectionLength = 2;
-        control.Paste();
-        Assert.Equal("bcxt", control.Text);
-        Assert.True(control.CanUndo);
-        Assert.True(control.Modified);
-        Assert.True(control.IsHandleCreated);
-        Assert.Equal(0, invalidatedCallCount);
-        Assert.Equal(0, styleChangedCallCount);
-        Assert.Equal(0, createdCallCount);
-    }
-
     [WinFormsFact]
     public void TextBoxBase_CreateHandle_Invoke_Success()
     {
@@ -4310,28 +4180,6 @@ public partial class TextBoxBaseTests
     }
 
     [WinFormsFact]
-    public void TextBoxBase_Cut_PasteNotEmpty_Success()
-    {
-        using SubTextBox control = new()
-        {
-            Text = "abc",
-            SelectionStart = 1,
-            SelectionLength = 2
-        };
-        control.Cut();
-        Assert.Equal("a", control.Text);
-        Assert.True(control.IsHandleCreated);
-
-        control.Text = "text";
-        control.SelectionLength = 2;
-        control.Paste();
-        Assert.Equal("bcxt", control.Text);
-        Assert.True(control.CanUndo);
-        Assert.True(control.Modified);
-        Assert.True(control.IsHandleCreated);
-    }
-
-    [WinFormsFact]
     public void TextBoxBase_Cut_PasteEmptyWithHandle_Success()
     {
         using SubTextBox control = new();
@@ -4353,42 +4201,6 @@ public partial class TextBoxBaseTests
         control.Text = "text";
         control.SelectionLength = 2;
         Assert.Equal("text", control.Text);
-        Assert.True(control.IsHandleCreated);
-        Assert.Equal(0, invalidatedCallCount);
-        Assert.Equal(0, styleChangedCallCount);
-        Assert.Equal(0, createdCallCount);
-    }
-
-    [WinFormsFact]
-    public void TextBoxBase_Cut_PasteNotEmptyWithHandle_Success()
-    {
-        using SubTextBox control = new()
-        {
-            Text = "abc",
-            SelectionStart = 1,
-            SelectionLength = 2
-        };
-        Assert.NotEqual(IntPtr.Zero, control.Handle);
-        int invalidatedCallCount = 0;
-        control.Invalidated += (sender, e) => invalidatedCallCount++;
-        int styleChangedCallCount = 0;
-        control.StyleChanged += (sender, e) => styleChangedCallCount++;
-        int createdCallCount = 0;
-        control.HandleCreated += (sender, e) => createdCallCount++;
-
-        control.Cut();
-        Assert.Equal("a", control.Text);
-        Assert.True(control.IsHandleCreated);
-        Assert.Equal(0, invalidatedCallCount);
-        Assert.Equal(0, styleChangedCallCount);
-        Assert.Equal(0, createdCallCount);
-
-        control.Text = "text";
-        control.SelectionLength = 2;
-        control.Paste();
-        Assert.Equal("bcxt", control.Text);
-        Assert.True(control.CanUndo);
-        Assert.True(control.Modified);
         Assert.True(control.IsHandleCreated);
         Assert.Equal(0, invalidatedCallCount);
         Assert.Equal(0, styleChangedCallCount);
@@ -5987,29 +5799,6 @@ public partial class TextBoxBaseTests
         Assert.False(control.IsHandleCreated);
     }
 
-    [WinFormsFact]
-    public void TextBoxBase_Paste_InvokeEmpty_Success()
-    {
-        using SubTextBox control = new();
-        control.Paste();
-        Assert.NotNull(control.Text);
-        Assert.True(control.IsHandleCreated);
-    }
-
-    [WinFormsFact]
-    public void TextBoxBase_Paste_InvokeNotEmpty_Success()
-    {
-        using SubTextBox control = new()
-        {
-            Text = "abc",
-            SelectionStart = 1,
-            SelectionLength = 2
-        };
-        control.Paste();
-        Assert.Equal("abc", control.Text);
-        Assert.True(control.IsHandleCreated);
-    }
-
     public static IEnumerable<object[]> ProcessCmdKey_TestData()
     {
         foreach (bool shortcutsEnabled in new bool[] { true, false })
@@ -7136,29 +6925,6 @@ public partial class TextBoxBaseTests
         Assert.Equal(0, invalidatedCallCount);
         Assert.Equal(0, styleChangedCallCount);
         Assert.Equal(0, createdCallCount);
-    }
-
-    [ActiveIssue("https://github.com/dotnet/winforms/issues/11559")]
-    [WinFormsFact]
-    [SkipOnArchitecture(TestArchitectures.X86,
-        "Flaky tests, see: https://github.com/dotnet/winforms/issues/11559")]
-    public void TextBoxBase_Undo_CanUndo_Success()
-    {
-        using SubTextBox control = new()
-        {
-            Text = "abc",
-            SelectionStart = 1,
-            SelectionLength = 2
-        };
-        control.Copy();
-
-        control.Text = "text";
-        control.SelectionLength = 2;
-        control.Paste();
-        Assert.Equal("bcxt", control.Text);
-
-        control.Undo();
-        Assert.Equal("text", control.Text);
     }
 
     public static IEnumerable<object[]> WndProc_ContextMenuWithoutContextMenuStrip_TestData()
